fix: refactored with creating new folders
