+++ handlers/handler.go
@@ -1,0 +1,209 @@
+package handlers
+
+import (
+	"context"
+	"fmt"
+	"net/http"
+	"time"
+
+	"github.com/ecoarchie/gin-restapi-recipes/models"
+	"github.com/gin-gonic/gin"
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/bson/primitive"
+	"go.mongodb.org/mongo-driver/mongo"
+)
+
+type RecipesHandler struct {
+	collection *mongo.Collection
+	ctx        context.Context
+}
+
+func NewRecipesHandler(ctx context.Context, collection *mongo.Collection) *RecipesHandler {
+	return &RecipesHandler{
+		collection: collection,
+		ctx:        ctx,
+	}
+}
+
+// swagger:operation GET /recipes recipes listRecipes
+// Returns list of recipes
+// ---
+// produces:
+// - application/json
+// responses:
+//
+//	'200':
+//		description: Successful operation
+func (h *RecipesHandler) ListRecipesHanlder(c *gin.Context) {
+	cursor, err := h.collection.Find(h.ctx, bson.M{})
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		fmt.Println("here")
+		return
+	}
+	defer cursor.Close(h.ctx)
+
+	recipes := make([]models.Recipe, 0)
+	for cursor.Next(h.ctx) {
+		var recipe models.Recipe
+		cursor.Decode(&recipe)
+		recipes = append(recipes, recipe)
+	}
+
+	c.JSON(http.StatusOK, recipes)
+}
+
+// swagger:operation POST /recipes recipes createRecipe
+// Create new recipe
+//
+// ---
+// produces:
+// - application/json
+// responses:
+//
+//	'200':
+//	  description: Recipe created
+//	'400':
+//	  description: Invalid input
+func (h *RecipesHandler) CreateRecipeHandler(c *gin.Context) {
+	var recipe models.Recipe
+	if err := c.ShouldBindJSON(&recipe); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": err.Error(),
+		})
+		return
+	}
+	recipe.ID = primitive.NewObjectID()
+	recipe.PublishedAt = time.Now()
+	_, err := h.collection.InsertOne(h.ctx, recipe)
+	if err != nil {
+		fmt.Println(err)
+		c.JSON(http.StatusInternalServerError, gin.H{
+			"error": "Error while inserting a new recipe",
+		})
+		return
+	}
+	c.JSON(http.StatusOK, recipe)
+}
+
+// swagger:operation PUT /recipes/{id} recipes updateRecipe
+// Update an existing recipe
+//
+// ---
+// produces:
+// - application/json
+// parameters:
+//   - name: id
+//     in: path
+//     description: ID of the recipe
+//     required: true
+//     type: string
+//
+// responses:
+//
+//	'200':
+//	  description: Successful operation
+//	'400':
+//	  description: Invalid input
+//	'404':
+//	  description: Invalid recipe ID
+func (h *RecipesHandler) UpdateRecipeHandler(c *gin.Context) {
+	id := c.Param("id")
+	var recipe models.Recipe
+	if err := c.ShouldBindJSON(&recipe); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": err.Error(),
+		})
+		return
+	}
+	objectID, _ := primitive.ObjectIDFromHex(id)
+	_, err := h.collection.UpdateOne(h.ctx, bson.M{
+		"_id": objectID,
+	}, bson.D{
+		{"$set", bson.D{
+			{"name", recipe.Name},
+			{"instructions", recipe.Instructions},
+			{"ingredients", recipe.Ingredients},
+			{"tags", recipe.Tags},
+		}},
+	})
+	if err != nil {
+		fmt.Println(err)
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+	c.JSON(http.StatusOK, gin.H{"message": "Recipe has been updated"})
+}
+
+// swagger:operation DELETE /recipes/{id} recipes deleteRecipe
+// Delete recipe by ID
+// ---
+// produces:
+// - application/json
+// parameters:
+//   - name: id
+//     in: path
+//     description: ID of the recipe
+//     required: true
+//     type: string
+//
+// responses:
+//
+//	'200':
+//	  description: recipe has been deleted
+//	'404':
+//	  description: recipe not found
+func (h *RecipesHandler) DeleteRecipeHandler(c *gin.Context) {
+	id := c.Param("id")
+	objectID, _ := primitive.ObjectIDFromHex(id)
+	_, err := h.collection.DeleteOne(h.ctx, bson.M{
+		"_id": objectID,
+	})
+	if err != nil {
+		fmt.Println(err)
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+	c.JSON(http.StatusOK, gin.H{
+		"message": "Recipe has been deleted",
+	})
+}
+
+// swagger:operation GET /recipes/search recipes searchRecipes
+// Search recipes by tag
+// ---
+// produces:
+// - application/json
+// parameters:
+//   - name: tag
+//     in: query
+//     description: tag to search for in recipes
+//     required: true
+//     type: string
+//
+// responses:
+//
+//	'200':
+//	  description: recipes found
+func (h *RecipesHandler) SearchRecipesHandler(c *gin.Context) {
+	tag := c.Query("tag")
+	listOfRecipes := make([]models.Recipe, 0)
+
+	cursor, err := h.collection.Find(h.ctx, bson.M{
+		"tags": tag,
+	})
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		fmt.Println("here")
+		return
+	}
+	defer cursor.Close(h.ctx)
+
+	for cursor.Next(h.ctx) {
+		var recipe models.Recipe
+		cursor.Decode(&recipe)
+		listOfRecipes = append(listOfRecipes, recipe)
+	}
+
+	c.JSON(http.StatusOK, listOfRecipes)
+}
